[Security Solution][Entity details] - move managed user code to flyout folder
