+++ x-pack/plugins/security_solution/public/flyout/entity_details/host_right/content.stories.tsx
@@ -10,7 +10,7 @@ import { storiesOf } from '@storybook/react';
 import { EuiFlyout } from '@elastic/eui';
 import { TestProvider } from '@kbn/expandable-flyout/src/test/provider';
 import { StorybookProviders } from '../../../common/mock/storybook_providers';
-import { mockRiskScoreState } from '../../../timelines/components/side_panel/new_user_detail/__mocks__';
+import { mockRiskScoreState } from '../../shared/mocks';
 import { HostPanelContent } from './content';
 import { mockObservedHostData } from '../mocks';
