+++ x-pack/plugins/security_solution/public/flyout/entity_details/user_right/content.tsx
@@ -14,8 +14,8 @@ import { AssetCriticalityAccordion } from '../../../entity_analytics/components/
 import { OBSERVED_USER_QUERY_ID } from '../../../explore/users/containers/users/observed_details';
 import { FlyoutRiskSummary } from '../../../entity_analytics/components/risk_summary_flyout/risk_summary';
 import type { RiskScoreState } from '../../../entity_analytics/api/hooks/use_risk_score';
-import { ManagedUser } from '../../../timelines/components/side_panel/new_user_detail/managed_user';
-import type { ManagedUserData } from '../../../timelines/components/side_panel/new_user_detail/types';
+import { ManagedUser } from './components/managed_user';
+import type { ManagedUserData } from './types';
 import type { RiskScoreEntity, UserItem } from '../../../../common/search_strategy';
 import { USER_PANEL_RISK_SCORE_QUERY_ID } from '.';
 import { FlyoutBody } from '../../shared/components/flyout_body';
